fix: docspage migration
