+++ package.json
@@ -23,7 +23,7 @@
     "minimatch": "^5.1.2",
     "next": "13.4",
     "next-image-export-optimizer": "^1.0.1",
-    "next-mdx-remote": "^4.1.0",
+    "next-mdx-remote": "^4.4.1",
     "next-remote-watch": "^2.0.0",
     "prism-react-renderer": "^1.3.5",
     "prismjs": "^1.29.0",
